resolving pre-fetch with edit-booking-data service file and with edit-booking-load component

+++ src/app/admin/rooms/room-detail/room-detail.component.ts
@@ -2,7 +2,6 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Room} from "../../../model/Room";
 import {Router} from "@angular/router";
 import {DataService} from "../../../data.service";
-import {error} from "@angular/compiler-cli/src/transformers/util";
 
 @Component({
   selector: 'app-room-detail',
@@ -30,14 +29,17 @@ export class RoomDetailComponent implements OnInit {
   }
 
   deleteRoom() {
-    this.message = 'deleting...';
-    this.dataService.deleteRoom(this.room.id).subscribe(
-      next => {
-        this.dataChangedEvent.emit();
-        this.router.navigate(['admin', 'rooms']);
-      }, error => {
-        this.message = 'Sorry - this room cannot be deleted at this time.'
-      }
-    );
+    const result = confirm('Are you sure you wish to delete this room?');
+    if (result) {
+      this.message = 'deleting...';
+      this.dataService.deleteRoom(this.room.id).subscribe(
+        next => {
+          this.dataChangedEvent.emit();
+          this.router.navigate(['admin', 'rooms']);
+        }, error => {
+          this.message = 'Sorry - this room cannot be deleted at this time.'
+        }
+      );
+    }
   }
 }
